requirements.txt
pyusb>=1.2.0
cryptography>=41.0
Pillow>=10.0
